Add files via upload
Sprite sheet.  I was thinking to keep the dungeon kinda dark and mysterious, but have items bright and stand out. kinda like the "happy" part of the dream. The green on some of the sprites is the color i want to make transparent so we can have some black outlines. ill upload this on github too
finalsheet.png