Added bosses and animated the fish
Bosses are now able to spawn, the fish is animated. However, I would like to change animation so that it works when the boss is moving around. Will work on that later.

The boss room may have to be a different floor color... bc my designs kinda blend in right now.

Updated the readme cuz why not.

Spritesheet is 100% up to date.